Enum fdc_gun_comms::FdcGunMessage

pub enum FdcGunMessage {
    StatusRequest,
    StatusReply {
        status: Status,
        rounds: HashMap<Ammunition, u32>,
    },
    FireReport {
        shot: u8,
        total_shots: u8,
        ammunition: Ammunition,
        target_location: TargetLocation,
        time_to_target: u32,
    },
    FireCommand {
        rounds: u32,
        ammunition: Ammunition,
        target_location: TargetLocation,
    },
    CheckFire,
    ComplianceResponse {
        compliance: Compliance,
    },
}

High-level message definition.

Intended to be used as an intermediate between raw bytes and a specific strongly typed message

Variants

StatusRequest

A request for status

StatusReply

Fields

status: Status

High-level status

rounds: HashMap<Ammunition, u32>

Rounds available

A reply for a status request

FireReport

Fields

shot: u8

The shot number for this report

total_shots: u8

The total number of shots that the gun issuing the report will fire

ammunition: Ammunition

The Ammunition in use

target_location: TargetLocation

The location of this shot

time_to_target: u32

The time in seconds from the issue of this FdcGunMessage::FireReport until the round is expected to land

A Report of gun fires

FireCommand

Fields

rounds: u32

Number of rounds to fire

ammunition: Ammunition

Type of ammunition for fires

target_location: TargetLocation

Location of target

A Command from the FDC to fire

CheckFire

A Check Fire command to a gun, to stop any active fires

ComplianceResponse

Fields

compliance: Compliance

The specific compliance type

A Compliance response to a CheckFire, FireCommand

Implementations

impl FdcGunMessage

pub fn serialize(&self, buf: &mut impl BufMut) -> Result<()>

Serializes an FdcGunMessage to the supplied buffer.

Errors

This method returns a std::io::Error if there is data missing, or if any data is otherwise invalid

pub fn deserialize(buf: impl Buf) -> Result<Self>

Deserializes an FdcGunMessage from the supplied buffer.

Errors

This method returns a std::io::Error if there is data missing, or if any data is otherwise invalid